+++ lib/qiita/markdown/filters/mention.rb
@@ -6,7 +6,7 @@ module Qiita
       #
       # You can pass :allowed_usernames context to limit mentioned usernames.
       class Mention < HTML::Pipeline::MentionFilter
-        MentionPattern = /
+        MENTION_PATTERN = /
           (?:^|\W)
           @((?>[\w][\w-]{1,30}\w(?:@github)?))
           (?!\/)
@@ -18,9 +18,14 @@ module Qiita
           )
         /ix
 
+        def call
+          restore_wrongly_emphasized_mentions
+          super
+        end
+
         # @note Override to use customized MentionPattern and allowed_usernames logic.
         def mention_link_filter(text, _, _)
-          text.gsub(MentionPattern) do |match|
+          text.gsub(MENTION_PATTERN) do |match|
             name = $1
             if allowed_usernames && !allowed_usernames.include?(name)
               match
@@ -36,6 +41,22 @@ module Qiita
         end
 
         private
+
+        # Given a user @foo_ and the following markdown:
+        #
+        #   _symbol @foo_
+        #
+        # The `@foo_` should be treated as a mention but Redcarpet parses it as an emphasis,
+        # so we restore the original source here.
+        def restore_wrongly_emphasized_mentions
+          doc.search("em").each do |node|
+            last_child = node.children.last
+            next if !last_child.text? || !last_child.text.match(MENTION_PATTERN)
+            node.prepend_child("_")
+            node.add_child("_")
+            node.replace(node.children)
+          end
+        end
 
         def allowed_usernames
           context[:allowed_usernames]